SC2-46 SC2-50 #done finish abstract logging functionality
