+++ Pipelines/src/main/java/pipelines/examples/SampleLoggingPipeline.java
@@ -1,0 +1,15 @@
+package pipelines.examples;
+
+import util.LoggingPipeline;
+import util.annotation.Disabled;
+import util.annotation.Pipeline;
+
+@Pipeline(camera=0)
+@Disabled
+public class SampleLoggingPipeline extends LoggingPipeline {
+
+    public SampleLoggingPipeline() {
+        super("sample", "raw-image");
+    }
+    
+}
